Add files (questions, student & teacher view)
Added three new files: AddQuestions.java, StudentView.java and TeacherView.java.

StudentView: this is the view that the students will see when they log into the system
TeacherView: this is the view that the teacher sees. This view allows teachers to add, view, edit, and remove. This view also includes a table.
AddQuestions: When the teacher presses the "add" button in the teacherview, the window that pops up is the AddQuestion window. This window allows the teacher to add a new question.