Add Login and CreateAccount UI
Uploaded CreateAccount.java and Login.java page. Both pages have a "Teacher" and a "Student" checkbox. If the user tries to register or login without entering a username or password, they will receive a message informing them of this. If the user logins as a student, they will be brought to the student view. If the user logins as a teacher, they will be brought in as a teacher. The "CreateAccount" page has a back button so that the user can go back to login page.